Creating RTypeSocketType enum that is  UDP or TCP to create a template class RTypeSocket you can use like that RTypeSocket<UDP|TCP> :) Remove CMake in Socket folder, now it's hpp only because template. No more LibrarySocket
