+++ Shared/Socket/Enum/RTypeSocketType.h
@@ -1,0 +1,13 @@
+//
+// Created by pasteu_e on 02/12/16.
+//
+
+#ifndef R_TYPE_RTYPESOCKETTYPE_H
+#define R_TYPE_RTYPESOCKETTYPE_H
+
+enum SocketType {
+        UDP = 0,
+        TCP = 1
+};
+
+#endif //R_TYPE_RTYPESOCKETTYPE_H
