+++ Shared/UnitTests/CMakeLists.txt
@@ -10,7 +10,7 @@ ELSE()
     set(LIBRARY_SOCKET_HEADER_FOLDER ../Socket/Linux)
 ENDIF()
 
-include_directories(../Include ../Socket/Includes ../Socket/Interfaces ${LIBRARY_SOCKET_HEADER_FOLDER})
+include_directories(../Include ../Socket/Includes ../Socket/Interfaces ../Socket/Enum ${LIBRARY_SOCKET_HEADER_FOLDER})
 
 foreach (_file ${SHARED_TEST_FILES})
     get_filename_component(SHARED_TEST_FILES_ABS ${_file} REALPATH)
